% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_corr.R
\name{bw_corr}
\alias{bw_corr}
\title{plot correlation scatterplot for bw files}
\usage{
bw_corr(dat, multiBigwigSummary_output = TRUE, pattern = NULL)
}
\arguments{
\item{dat}{Absolute path of count matrix. The input file can be generated by
deeptools \code{multiBigwigSummary bins} command. The output should be a
counts matrix \code{--outRawCounts} with first three columns:\code{chr
start end} and later sample wise counts.  Alternatively, a matrix of
expression values with first column gene-name and later expression values.}

\item{multiBigwigSummary_output}{Logical, whether the input matrix is
\code{multiBigwigSummary bins} out put or not.\code{ default:TRUE}}

\item{pattern}{A character vector, defining pattern to split the columns in
replicates. \code{ default: E.g. sampleA_set1, sampleA_set2, sampleB_set1,
sampleB_set2. pattern="(.*)_(.*)". will have groups of sample: sampleA, sampleB
and replicate: set1, set2}}
}
\value{
a multi-sample scatterplot of log2 values with correlation value
}
\description{
correlation between two replicates of samples with correlation
  value.
}
\examples{
\dontrun{

dat <- read.delim("inst/extdata/correlation_data/an_histone_spore_count_mat.tab",sep="\\t", header = TRUE)
gp <- bw_corr(dat, pattern = ".*_spore_(.*)_(.*)")
ggsave(gp, filename = "An_spore_histone_bwscatterplot.png", height=5, width=5, device = "png", units = "in")
}
}
\author{
pooja sethiya
}
