% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_heatmap.R
\name{ggplot_heatmap}
\alias{ggplot_heatmap}
\title{heatmap of gene expressions by class}
\usage{
ggplot_heatmap(input_data, threshold = FALSE, output_name = "Sample")
}
\arguments{
\item{input_data}{A tibble, with four columns. First column name should be
\code{genes} and second column name should be \code{class}. Other columns
can have unique names to plot.}

\item{threshold}{Logical, to set maximum limit of the color
key.\code{default:FALSE}. If \code{TRUE} on-run threshold value should be
provided.}

\item{output_name}{A character vector, containing name of the output
plot.\code{default:Sample}}
}
\value{
heatmap of expression of given genes categorized by class and output
  image file of same.
}
\description{
Heatmap of genes expression value categorized by classes
}
\examples{
\dontrun{

dat <- readr::read_delim("data/an_spore_hypha_specificgenes.txt", delim="\\t", col_names = TRUE)
input_data <- dat \%>\% dplyr::filter(class=="spore_maturation")
ggplot_heatmap(input_data ,threshold = TRUE, output_name = "plots/An_spores_maturation_genes_exprsn")

}
}
\author{
pooja sethiya
}
