% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_diamond.R
\name{GO_diamondplot}
\alias{GO_diamondplot}
\title{GO terms in diamond plot}
\usage{
GO_diamondplot(data, palette = TRUE, output_name)
}
\arguments{
\item{data}{A tibble, with following columns \itemize{\item term: character,
GO term \item annotated: numeric, number of genes enriched in the GO-term
\item total: numeric, total number of genes assigned to the GO-term \item
pvalue: numeric, pvalue of enrichment \item category: character, a
group-name to annotate the GO and color the diamonds \item class:
character, optional if there are two categories, second category should be
under column class. When class is present then the diamonds are colored by
class else by category}}

\item{palette}{Character, to choose from available palette or define a
  palette.\code{default:TRUE} \itemize{ \item two_color: a palette of pink and green
\item three_color:  a palette of  orange, pink and green
\item NULL: user input of character vector of colors
\item TRUE: a palette of colors in rainbow
  }}

\item{output_name}{A character vector containing name of the output
plot.\code{default:Sample}}
}
\value{
a diamond GO plot and image file of same.
}
\description{
Visualize GO enrichment result in a diamond plot
}
\examples{
\dontrun{

data <- readr::read_delim("data/an_af_pm_spores_GO.txt", col_names = TRUE, delim="\\t")
GO_diamondplot(data, output_name = "plots/An_Af_Pm_GO", palette="three_color")

}
}
\author{
pooja sethiya
}
