% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot_from_bw.R
\name{lineplot_for_bw}
\alias{lineplot_for_bw}
\title{lineplot for bw file}
\usage{
lineplot_for_bw(feature_txDb, genelist, bw_file, output_name = "Sample",
  tss = "TRUE", expression_value = TRUE, tiles = 4, median = TRUE)
}
\arguments{
\item{feature_txDb}{A TxDb object of gene feature file, generated by
\code{GenomicFeatures::makeTxDbFromGFF()}}

\item{genelist}{A tibble with two columns \code{1. gene_id and 2.Expression
value or Percentile rank} without header, to subset genes from the feature
file.}

\item{output_name}{A character vector containing name of the output
plot.\code{default:Sample}}

\item{tss}{Logical, if TRUE plot 1kb up/downstream of start co-ordinate; if
FALSE plot 1kb up/downstream of gene-body \code{default: TRUE}}

\item{expression_value}{Logical, if TRUE percentile is calculated based on
the expression value column (ie. second column), if FALSE second column
should define the percentiles. \code{default: TRUE}}

\item{tiles}{Numeric, when \code{expression_value=TRUE}, specify the number
of percentiles inorder to divide the data.\code{default: 4}}

\item{median}{Logical, to display median or mean line of
percentiles.\code{default: TRUE}}

\item{bw_files}{A character vector containing either name or names of GRanges
object of bw or bdg files. GRanges object for bw_files can be generated by: \code{
rtracklayer::import()}}
}
\value{
A line-plot of percentiles and a tibble of percentiles appended to
  expression data.
}
\description{
Plot line-plot of each percentile for given gene-list.
}
\examples{

\dontrun{

 # Load feature file from system data, to run example data
 feature_txDb <- AnnotationDbi::loadDb(system.file("extdata/sqllite/an_feature_file_s10_m04_r07.sqlite" , package = "FungalSporeAnalysis"))

 # Or load feature file as TxDb
 feature_txDb <- GenomicFeatures::makeTxDbFromGFF("an_feature_file_s10_m04_r07.gff")

 # load list of genes to subset from all the genes
 genelist_dat <- system.file("extdata/genesets/an_spore_pol2_for_percentilelineplot.txt" , package = "FungalSporeAnalysis")
 genelist <- readr::read_delim(genelist_dat,delim="\\t", col_names=FALSE)

 # Load bw files as GRanges object
 pol2_veA_wt_spore <- rtracklayer::import.bw("pol2_veA_wt_spore_mix22_CACAGTTGGT_normalized_repeat.bw")

 # Load GRanges object from system data, to run example data
 data_files <- system.file("extdata/sysdata.rda" , package = "FungalSporeAnalysis")
 load(data_files)

 ###########################################

 lineplot_for_bw(feature_txDb, genelist =genelist,bw_file="pol2_veA_wt_spore", output_name = "pol2_veA_wt_spore")

}
}
\author{
pooja sethiya
}
