% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles_normalized_by_control.R
\name{profiles_normalized_by_control}
\alias{profiles_normalized_by_control}
\title{profile plots of signal normalized to control}
\usage{
profiles_normalized_by_control(feature_txDb, bw_test, bw_control,
  genelist_1 = NULL, genelist_2 = NULL, ymax = 6, ymin = 1,
  output_name = "Sample")
}
\arguments{
\item{feature_txDb}{A TxDb object of gene feature file, generated by
\code{GenomicFeatures::makeTxDbFromGFF()}}

\item{bw_test}{A character vector containing name of the bw file object.}

\item{bw_control}{A character vector containing name of the bw file object to
normalize the bw_test.}

\item{genelist_1}{A tibble with two columns \code{1. gene_id and 2.Expression
value} without header, to subset genes from the feature file.
\code{default:NULL}}

\item{genelist_2}{A tibble with either one or two columns to be considered as
control genes.  \code{1. gene_id and/or 2.Expression value} without header,
to subset genes from the feature file. \code{default:NULL}}

\item{ymax}{Numeric, y-axis max limit for lineplot \code{default:6}}

\item{ymin}{Numeric, y-axis minimum limit for lineplot \code{default:1}}

\item{output_name}{A character vector containing name of the output
plot.\code{default:Sample}}
}
\value{
An image of two heatmaps normalized to control bw. \itemize{
  \item\code{Heatmap_list1:} profile of genelist_1 ordered by expression
  values (high to low). \item\code{Heatmap_list1:} profile of genelist_2. }
}
\description{
Control normalized signal are plotted for all genes or specific list of
genes.
}
\examples{
\dontrun{
 feature_txDb <- AnnotationDbi::loadDb(system.file("extdata/sqllite/an_feature_file_s10_m04_r07.sqlite" , package = "FungalSporeAnalysis"))
 data_files <- system.file("extdata/sysdata.rda" , package = "FungalSporeAnalysis")

 genelist_1 <- readr::read_delim(system.file("extdata/genesets/an_spore_pol2.txt" , package = "FungalSporeAnalysis"), delim="\\t", col_names = FALSE)
 genelist_2 <- readr::read_delim(system.file("extdata/genesets/an_spore_pol2_controlgenes.txt", package = "FungalSporeAnalysis"), delim="\\t", col_names = FALSE)

 profiles_normalized_by_control(feature_txDb = feature_txDb,bw_test = "H3AC_veA_wt_spore", bw_control = "H3_an_spore",genelist_1 = genelist_1,genelist_2 = genelist_2, ymax = 3.8,ymin=0.5, output_name = "H3Ac_veA_wt_spore")

}
}
\author{
pooja sethiya
}
