% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genelist_specific_profileplot.R
\name{genelist_specific_profileplot}
\alias{genelist_specific_profileplot}
\title{profile plot specific to a genelist}
\usage{
genelist_specific_profileplot(feature_txDb, bw_files, genelist = NULL,
  tss = TRUE, palette = "cream_brown", max_key = 6, min_key = 0,
  top_line = TRUE, ymax = 6, ymin = 1, log2 = TRUE,
  rename = FALSE, output = TRUE, output_name = "Sample")
}
\arguments{
\item{feature_txDb}{A TxDb object of gene feature file, generated by
\code{GenomicFeatures::makeTxDbFromGFF()}}

\item{bw_files}{A character vector containing either name or names of GRanges
object of bw or bdg files. GRanges object for bw_files can be generated by: \code{
rtracklayer::import()}}

\item{genelist}{A tibble with two columns \code{1. gene_id and 2.Expression
value} without header, to subset genes from the feature file.
\code{default:NULL}}

\item{tss}{Logical, if TRUE plot 1kb up/downstream of start co-ordinate; if
FALSE plot 1kb up/downstream of gene-body \code{default: TRUE}}

\item{palette}{A character vector to choose sequential palette; \itemize{
\item white_green \item white_red \item white_blue \item cream_pink \item
cream_brown \item cream_green } \code{default: cream_brown}}

\item{max_key}{Numeric, maximum limit of the color key.\code{default:6}}

\item{min_key}{Numeric, mimimum limit of the color key.\code{default:0}}

\item{top_line}{Logical, whether to plot profile with average signal as line
plot at the top of plot. \code{default: TRUE}}

\item{ymax}{Numeric, y-axis maximum limit for lineplot \code{default:6}}

\item{ymin}{Numeric, y-axis minimum limit for lineplot \code{default:1}}

\item{log2}{Logical, whether to plot profile as log-transformed or not.}

\item{rename}{Logical, whether to remove characters after first \code{- OR
_}, suitable for long names of bw objects.\code{default: FALSE}}

\item{output}{Logical, to return a output of profile plot or not.
\code{default: TRUE, with output_name,if FALSE; plot will be returned} **
Highly recommended to plot in outfile when .bw files are more than 2.}

\item{output_name}{A character vector containing name of the output
plot.\code{default:Sample}}
}
\value{
Parallel plots of profiles(log2) for given gene lists, ordered by
  given expression value (high to low).
}
\description{
Profile plot for list of genes to be subsetted from genome
  feature file.
}
\examples{
\dontrun{

 # Load feature file from system data, to run example data
 feature_txDb <- AnnotationDbi::loadDb(system.file("extdata/sqllite/an_feature_file_s10_m04_r07.sqlite" , package = "FungalSporeAnalysis"))

 # Or load feature file as TxDb
 feature_txDb <- GenomicFeatures::makeTxDbFromGFF("an_feature_file_s10_m04_r07.gff")

 # Load bw files as GRanges object
 pol2_veA_wt_spore <- rtracklayer::import.bw("pol2_veA_wt_spore_mix22_CACAGTTGGT_normalized_repeat.bw")
 TBP_veA_wt_spore <- rtracklayer::import.bw("TBP_veA_wt_spore_mix22_CAGTTGGT_normalized_repeat.bw")

 # Load GRanges object from system data, to run example data
 data_files <- system.file("extdata/sysdata.rda" , package = "FungalSporeAnalysis")
 load(data_files)

 # Load example gene_list file
 gene_list <- readr::read_delim(system.file("extdata/genesets/an_spore_pol2.txt" , package = "FungalSporeAnalysis"), delim="\\t", col_names = FALSE)

 #################

 bw_files <- c("pol2_veA_wt_spore","TBP_veA_wt_spore")
 genelist_specific_profileplot(feature_txDb=feature_txDb,bw_files = bw_files, genelist=gene_list, output_name="An_Pol2_TBP_TFIIB", ymin=3)

}
}
\author{
pooja sethiya
}
