% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot_from_bw.R
\name{lineplot_for_bw}
\alias{lineplot_for_bw}
\title{lineplot for bw file}
\usage{
lineplot_for_bw(feature_txDb, genelist, bw_file, output_name = "Sample",
  tss = "TRUE", expression_value = TRUE, tiles = 4, median = TRUE)
}
\arguments{
\item{feature_txDb}{A TxDb object of gene feature file, generated by
\code{GenomicFeatures::makeTxDbFromGFF()}}

\item{genelist}{A tibble with two columns \code{1. gene_id and 2.Expression
value or Percentile rank} without header, to subset genes from the feature
file.}

\item{bw_file}{A character vector containing either name of the bw file
object or names of bw file objects to plot.}

\item{output_name}{A character vector containing name of the output
plot.\code{default:Sample}}

\item{tss}{Logical, if TRUE plot 1kb up/downstream of start co-ordinate; if
FALSE plot 1kb up/downstream of gene-body \code{default: TRUE}}

\item{expression_value}{Logical, if TRUE percentile is calculated based on
the expression value column (ie. second column), if FALSE second column
should define the percentiles. \code{default: TRUE}}

\item{tiles}{Numeric, when \code{expression_value=TRUE}, specify the number
of percentiles inorder to divide the data.\code{default: 4}}

\item{median}{Logical, to display median or mean line of percentiles.\code{default: TRUE}}
}
\value{
A line-plot of percentiles and a tibble of percentiles appended to expression data.
}
\description{
Plot line-plot of each percentile for given gene-list.
}
\examples{
\dontrun{

feature_txDb <- AnnotationDbi::loadDb("R/an_feature_file_s10_m04_r07.sqlite")
genelist <- readr::read_delim("data/an_spore_pol2_for_percentilelineplot.txt",delim="\\t", col_names=FALSE)
lineplot_for_bw(feature_txDb, genelist =genelist,bw_file="pol2_veA_wt_spore", output_name = "plots/pol2_veA_wt_spore")

}
}
\author{
pooja sethiya
}
